+++ src/DependencyInjection/ContainerProvider.php
@@ -6,8 +6,8 @@ use Jarvis\Skill\Core\CallbackResolver;
 use Jarvis\Skill\Core\ScopeManager;
 use Jarvis\Annotation\Parser;
 use Jarvis\Annotation\Handler\ResponseFormatHandler;
-use Jarvis\Event\JarvisEvents;
-use Jarvis\Event\Receiver\ControllerReceiver;
+use Jarvis\EventBroadcaster\JarvisEvents;
+use Jarvis\EventBroadcaster\Receiver\ControllerReceiver;
 use Jarvis\Jarvis;
 use Jarvis\Relational\Annotation\Handler\ParamConverterHandler;
 use Jarvis\Rest\EventReceiver\RestReceiver;
